Clean up Android scan details (#26)
* fix ios-only permission thing

* fix error blocking android dump

* fix package_info parse error on edge case

* make permissions tuples with description as the last time used

* revamp AndroidDump.info()

* clean parse dump further

* get install time android (did not test on scan)

* only show information we have

* get rid of permissions exception

* mostly comments and moving a try except block

+++ scripts/pull_apks.sh
@@ -33,7 +33,9 @@ ${adb_cmd} shell "mkdir -p /sdcard/apps/"
 function pull {
     echo "Clearing the folder"
     ${adb_cmd} pull "/sdcard/apps/" $PKG_DIR
-    mv "$PKG_DIR"/apps/* "$PKG_DIR"
+    if [[ -e "$PKG_DIR/apps/* " ]]; then
+        mv "$PKG_DIR"/apps/* "$PKG_DIR"
+    fi
     ${adb_cmd} shell "rm -rf /sdcard/apps/"
     ${adb_cmd} shell "mkdir -p /sdcard/apps/"
 }
